requirements.txt
langchain 
openai 
python-dotenv 
GitPython 
pinecone-client
streamlit
tiktoken